moved into folders, circuit performance improvements, added seed
