+++ CMakeLists.txt
@@ -38,6 +38,8 @@ if (DEFINED qilisim AND qilisim STREQUAL "ON")
     message(STATUS "[QiliSim] All dependencies found, preparing QiliSim build files")
     pybind11_add_module(qilisim_module src/qilisdk_cpp/backends/qilisim/qilisim.cpp)
     file(GLOB QILISIM_SOURCES src/qilisdk_cpp/backends/qilisim/*.cpp)
+    file(GLOB QILISIM_SUBDIR_SOURCES src/qilisdk_cpp/backends/qilisim/*/*.cpp)
+    list(APPEND QILISIM_SOURCES ${QILISIM_SUBDIR_SOURCES})
     target_sources(qilisim_module PRIVATE ${QILISIM_SOURCES})
 
     # Set compile options
