+++ tests/backends/test_qilisim_backend.py
@@ -61,8 +61,7 @@ def test_cnot(backend):
     assert result.samples == {"00": 10}
 
 
-def test_exponential_gates():
-    backend = QiliSim()
+def test_exponential_gates(backend):
     circuit = Circuit(nqubits=1)
     circuit.add(X(0).exponential())
     result = backend.execute(Sampling(circuit=circuit, nshots=100))
@@ -72,8 +71,7 @@ def test_exponential_gates():
     assert "0" in samples
 
 
-def test_nshots():
-    backend = QiliSim()
+def test_nshots(backend):
     circuit = Circuit(nqubits=1)
     result = backend.execute(Sampling(circuit=circuit, nshots=10))
     assert isinstance(result, SamplingResult)
@@ -109,10 +107,9 @@ def test_basic_gates(backend, gate):
     assert isinstance(result, SamplingResult)
 
 
-def test_multi_controlled_execution():
+def test_multi_controlled_execution(backend):
     # Create two Xs then a multi-controlled X (Toffoli) gate
     # Expect roughly all shots to be '111'
-    backend = QiliSim()
     circuit = Circuit(nqubits=3)
     circuit.add(X(0))
     circuit.add(X(1))
@@ -137,13 +134,12 @@ def test_multiple_parameterized_gates(backend):
 
 
 def test_many_gates(backend):
-    c = Circuit.random(nqubits=2, single_qubit_gates={H, X, Y, Z, T, RX, RZ}, two_qubit_gates={CNOT}, ngates=10000)
+    c = Circuit.random(nqubits=2, single_qubit_gates={H, X, Y, Z, T, RX, RZ}, two_qubit_gates={CNOT}, ngates=1000)
     result = backend.execute(Sampling(circuit=c, nshots=1000))
     assert isinstance(result, SamplingResult)
 
 
-def test_measurement_gates():
-    backend = QiliSim()
+def test_measurement_gates(backend):
     circuit = Circuit(nqubits=2)
     circuit.add(X(0))
     circuit.add(M(0))
@@ -154,7 +150,7 @@ def test_measurement_gates():
     assert samples["1"] == 50
 
 
-def test_constant_hamiltonian():
+def test_constant_hamiltonian(backend):
     x = 2.0
     schedule = Schedule(
         hamiltonians={"hz": x * pauli_z(0)},
@@ -164,7 +160,6 @@ def test_constant_hamiltonian():
     )
     psi0 = ket(0)
     obs = [pauli_z(0)]
-    backend = QiliSim()
     res = backend.execute(
         TimeEvolution(schedule=schedule, initial_state=psi0, observables=obs, store_intermediate_results=True)
     )
@@ -209,7 +204,7 @@ def test_time_dependent_hamiltonian(method):
     assert np.isclose(expect_z, -1.0, rtol=1e-2)
 
 
-def test_time_dependent_hamiltonian_qtensor_observable():
+def test_time_dependent_hamiltonian_qtensor_observable(backend):
     o = 1.0
     dt = 0.5
     T = 100
@@ -225,7 +220,6 @@ def test_time_dependent_hamiltonian_qtensor_observable():
         QTensor(pauli_z(0).to_matrix()),  # measure z as QTensor
     ]
 
-    backend = QiliSim()
     res = backend.execute(TimeEvolution(schedule=schedule, initial_state=psi0, observables=obs))
 
     assert isinstance(res, TimeEvolutionResult)
@@ -235,7 +229,7 @@ def test_time_dependent_hamiltonian_qtensor_observable():
     assert np.isclose(expect_z, -1.0, rtol=1e-2)
 
 
-def test_time_dependent_hamiltonian_pauli_observable():
+def test_time_dependent_hamiltonian_pauli_observable(backend):
     o = 1.0
     dt = 0.5
     T = 100
@@ -252,7 +246,6 @@ def test_time_dependent_hamiltonian_pauli_observable():
         pauli_z_pauli(0),
     ]
 
-    backend = QiliSim()
     res = backend.execute(TimeEvolution(schedule=schedule, initial_state=psi0, observables=obs))
 
     assert isinstance(res, TimeEvolutionResult)
@@ -262,7 +255,7 @@ def test_time_dependent_hamiltonian_pauli_observable():
     assert np.isclose(expect_z, -1.0, rtol=1e-2)
 
 
-def test_time_dependent_hamiltonian_bad_observable():
+def test_time_dependent_hamiltonian_bad_observable(backend):
     o = 1.0
     dt = 0.5
     T = 100
@@ -278,7 +271,6 @@ def test_time_dependent_hamiltonian_bad_observable():
         3.5,  # invalid observable
     ]
 
-    backend = QiliSim()
     with pytest.raises(ValueError, match=r"Observable type not recognized."):
         backend.execute(TimeEvolution(schedule=schedule, initial_state=psi0, observables=obs))
 
@@ -406,7 +398,7 @@ def test_monte_carlo_time_evolution(method):
 
     expect_z = res.final_expected_values[0]
     assert res.final_state.shape == (2, 2)
-    assert np.isclose(expect_z, -0.8, rtol=1e-2)
+    assert np.isclose(expect_z, -0.8, rtol=1e-1)
 
 
 def test_qilisim_params():
